secmet: style pass
